requirements.txt
pytest~=8.4.2
selenium~=4.36.0
pytest-html
allure-pytest
allure-python-commons